feat: access adapter removal bigint (#2369)
Resolves INT-5728, INT-5727, INT-5729.

Summary of changes:

- Everywhere the code used `Rv32HeapAdapterAir`, we switch to use
`Rv32VecHeapAdapterAir`.
- Everywhere the code used `Rv32HeapBranchAdapterAir`, we switch to use
a new `Rv32HeapBranchAdapterAir`, which accesses memory in the same way
as `Rv32VecHeapAdapterAir`, but is compatible with the branch CoreAirs.
- No other code uses `Rv32HeapAdapterAir` and
`Rv32HeapBranchAdapterAir`, so the `heap.rs` and `heap_branch.rs` files
were deleted.
- The interface for `Rv32VecHeapAdapterAir` and
`Rv32HeapBranchAdapterAir ` now becomes different to what the CoreAirs
expect, so wrappers in `vec_to_heap.rs` are used to convert between
them.

---------

+++ crates/vm/src/arch/testing/cuda.rs
@@ -210,7 +210,8 @@ impl TestBuilder<F> for GpuChipTestBuilder {
     ) -> (usize, usize) {
         let register = self.get_default_register(reg_increment);
         let pointer = self.get_default_pointer(pointer_increment);
-        self.write(1, register, pointer.to_le_bytes().map(F::from_u8));
+        // Cast to u32 to ensure we write exactly 4 bytes (RV32 register size).
+        self.write(1, register, (pointer as u32).to_le_bytes().map(F::from_u8));
         (register, pointer)
     }
 
@@ -368,7 +369,12 @@ impl GpuChipTestBuilder {
         pointer: usize,
         writes: Vec<[F; NUM_LIMBS]>,
     ) {
-        self.write(1usize, register, pointer.to_le_bytes().map(F::from_u8));
+        // Cast to u32 to ensure we write exactly 4 bytes (RV32 register size).
+        self.write(
+            1usize,
+            register,
+            (pointer as u32).to_le_bytes().map(F::from_u8),
+        );
         if NUM_LIMBS.is_power_of_two() {
             for (i, &write) in writes.iter().enumerate() {
                 self.write(2usize, pointer + i * NUM_LIMBS, write);
